refactor(IssueQueryController): rename package
