+++ backend/src/test/java/elbin_bank/issue_tracker/issue/presentation/IssueQueryControllerTest.java
@@ -2,6 +2,7 @@ package elbin_bank.issue_tracker.issue.presentation;
 
 import elbin_bank.issue_tracker.issue.application.query.IssueQueryService;
 import elbin_bank.issue_tracker.issue.application.query.dto.IssuesResponseDto;
+import elbin_bank.issue_tracker.issue.presentation.query.IssueQueryController;
 import elbin_bank.issue_tracker.label.infrastructure.query.projection.LabelProjection;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
